extract sprite.h
